requirements.txt
qrcode[pil]
flask
mysql-connector-python